backend/requirements.txt
fastapi
uvicorn 
passlib[bcrypt] 
python-jose
bcrypt==4.0.1 
dotenv
supabase
python-multipart
sentence-transformers
python-socketio